# docs/phase0/openapi-lmn-authority-v1.yaml
openapi: 3.1.0
info:
  title: LMN Authority API
  version: 1.0.0
  summary: linuxmuster.net LINBO Host & Config Authority API
  description: |
    The LMN Authority API runs on the linuxmuster.net server (10.0.0.11) and
    serves as the **single source of truth** for LINBO host registrations,
    start.conf configurations, and DHCP reservations.

    **LINBO Docker consumes this API read-only.** Docker never writes host,
    config, or room data back to the LMN server. All CRUD for these entities
    happens on the LMN side (webui7 / linuxmuster-import-devices).

    The LINBO Runtime Docker stack polls this API via a cursor-based delta feed,
    fetches changed records in batches, and builds local snapshots that are
    served to PXE clients over TFTP / HTTP / rsync.

    ## Authentication
    Every request requires a **Bearer token** in the `Authorization` header.
    Additionally, the caller's IP address must be in the server's allowlist
    (typically `10.0.0.0/16`).

    ## Cursor Format
    The delta-feed cursor is an opaque string with the format
    `"<unix-timestamp>:<sequence>"`, e.g. `"1708943200:42"`.
    Cursors are monotonically increasing and safe to compare lexicographically.
    An empty or missing cursor triggers a full-snapshot response (all IDs).

    ## Rate Limits
    The API enforces a per-token rate limit of **60 requests / minute**.
    Batch endpoints count as a single request regardless of payload size.
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/linbo/changes:
    get:
      operationId: getChanges
      tags: [delta]
      summary: Cursor-based delta feed
      description: |
        Returns a list of entity IDs that changed since the provided cursor.
        The LINBO Runtime Docker calls this endpoint on a polling interval
        (e.g. every 30 s) to discover what needs to be re-fetched.

        **Initial sync:** Pass an empty string `""` for `since` to receive a
        full snapshot containing all host MACs, all start.conf IDs, and all
        config IDs. The response includes `nextCursor` which must be stored
        and sent in subsequent requests.

        **Incremental sync:** Pass the last `nextCursor` value. The response
        contains only the entities that changed since that cursor.

        Changes are **rare** (a few times per day for ~2000 hosts / ~20
        configs), so most incremental responses will be empty.
      parameters:
        - name: since
          in: query
          required: true
          description: |
            Cursor from a previous response. Pass an empty string `""` for
            the initial full-snapshot request (returns all IDs).
            Must always be provided (required).
          schema:
            type: string
          examples:
            initial:
              summary: Initial sync (empty cursor)
              value: ""
            incremental:
              summary: Incremental sync
              value: "1708943200:42"
      responses:
        "200":
          description: Change set since the given cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeltaResponse"
              examples:
                empty:
                  summary: No changes since last poll
                  value:
                    nextCursor: "1708943200:42"
                    hostsChanged: []
                    startConfsChanged: []
                    configsChanged: []
                    dhcpChanged: false
                    deletedHosts: []
                    deletedStartConfs: []
                incremental:
                  summary: A host MAC changed, DHCP stale
                  value:
                    nextCursor: "1708943260:43"
                    hostsChanged:
                      - "AA:BB:CC:DD:EE:01"
                      - "AA:BB:CC:DD:EE:02"
                    startConfsChanged:
                      - "win11_efi_sata"
                    configsChanged: []
                    dhcpChanged: true
                    deletedHosts: []
                    deletedStartConfs: []
                fullSnapshot:
                  summary: Full snapshot (empty cursor)
                  value:
                    nextCursor: "1708943200:42"
                    hostsChanged:
                      - "AA:BB:CC:DD:EE:01"
                      - "AA:BB:CC:DD:EE:02"
                      - "AA:BB:CC:DD:EE:03"
                      - "11:22:33:44:55:66"
                    startConfsChanged:
                      - "win11_efi_sata"
                      - "ubuntu_efi_nvme"
                      - "mint_efi_sata"
                    configsChanged:
                      - "win11_efi_sata"
                      - "ubuntu_efi_nvme"
                      - "mint_efi_sata"
                    dhcpChanged: true
                    deletedHosts: []
                    deletedStartConfs: []
        "400":
          description: Invalid cursor format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "INVALID_CURSOR"
                message: "Cursor must be in format 'timestamp:sequence' or empty for full snapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"

  # =========================================================================
  # Batch Fetch — Hosts
  # =========================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
      description: |
        Pre-shared API token configured on the linuxmuster.net server.
        Additionally, the caller's IP must be in the server's allowlist
        (default: `10.0.0.0/16`).

  # =========================================================================
  # Reusable Responses
  # =========================================================================
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "UNAUTHORIZED"
            message: "Missing or invalid Authorization header"
    Forbidden:
      description: Caller IP not in allowlist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "FORBIDDEN"
            message: "Source IP 192.168.1.50 is not in the allowlist"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the rate limit resets
          example: 30
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "RATE_LIMITED"
            message: "Rate limit exceeded. Retry after 30 seconds."

  # =========================================================================
  # Schemas
  # =========================================================================
  schemas:
...
    DeltaResponse:
      type: object
      required:
        - nextCursor
        - hostsChanged
        - startConfsChanged
        - configsChanged
        - dhcpChanged
        - deletedHosts
        - deletedStartConfs
      properties:
        nextCursor:
          type: string
          description: |
            Opaque cursor to pass as `since` in the next poll.
            Format: `"<unix-timestamp>:<sequence>"`.
          examples: ["1708943260:43"]
        hostsChanged:
          type: array
          items:
            type: string
            pattern: "^([0-9A-Fa-f]{2}:){5}[0-9A-Fa-f]{2}$"
          description: MAC addresses of hosts that were created or updated
        startConfsChanged:
          type: array
          items:
            type: string
          description: IDs of start.conf records that were created or updated
        configsChanged:
          type: array
          items:
            type: string
          description: IDs of parsed config records that were created or updated
        dhcpChanged:
          type: boolean
          description: |
            `true` if any change affects DHCP exports (host IP, MAC, PXE flag,
            or hostgroup changed). The consumer should re-fetch DHCP exports.
        deletedHosts:
          type: array
          items:
            type: string
            pattern: "^([0-9A-Fa-f]{2}:){5}[0-9A-Fa-f]{2}$"
          description: MAC addresses of hosts that were deleted since the cursor
        deletedStartConfs:
          type: array
          items:
            type: string
          description: IDs of start.conf records that were deleted since the cursor

    # -----------------------------------------------------------------------
    # Host
    # -----------------------------------------------------------------------
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
          examples: ["UNAUTHORIZED"]
        message:
          type: string
          description: Human-readable error description
          examples: ["Missing or invalid Authorization header"]
        details:
          type: object
          additionalProperties: true
          description: Optional structured error details